+++ src/main/java/eu/unicore/uftp/standalone/oidc/OIDCAgentProxy.java
@@ -9,9 +9,7 @@ import java.nio.channels.SocketChannel;
 
 /**
  * Connector to the 'oidc-agent' via UNIX domain socket.
- * 
- * Modeled after the JSch ssh-agent proxy
- *  
+ *
  * @author schuller
  */
 public class OIDCAgentProxy {